% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTimeAnalysisByEpisode.R
\name{CreateTimeAnalysisByEpisodes}
\alias{CreateTimeAnalysisByEpisodes}
\title{CreateTimeAnalysisByEpisodes}
\usage{
CreateTimeAnalysisByEpisodes(
  HRVData,
  Tag = "",
  size = 60,
  doOutOfEpisodeAnalysis = FALSE,
  numofbins = NULL,
  interval = 7.8125,
  verbose = NULL
)
}
\arguments{
\item{HRVData}{Data structure that stores the beats register and information related to it.}

\item{Tag}{Type of episode}

\item{size}{Length of the window by window analysis - default is 300}

\item{doOutOfEpisodeAnalysis}{Boolean to do out of episodes analysis}

\item{numofbins}{Number of bins used in the analysis - If null it is estimated}

\item{interval}{Time interval used in the analysis}

\item{verbose}{Deprecated argument maintained for compatibility, use SetVerbose() instead}
}
\value{
Returns a list with two objects, that is, the values of the application of the selected function
 inside ("resultIn") and outside ("resultOut") episodes
}
\description{
Creates time analysis by episode.
 .
}
